apply new functionality off home, remove some files, to prevent problems in a future
