--- decidim-home/app/helpers/decidim/home/application_helper.rb
@@ -1,6 +1,0 @@
-module Decidim
-  module Home
-    module ApplicationHelper
-    end
-  end
-end
