--- decidim-home/app/jobs/decidim/home/application_job.rb
@@ -1,6 +1,0 @@
-module Decidim
-  module Home
-    class ApplicationJob < ActiveJob::Base
-    end
-  end
-end
