--- decidim-home/app/mailers/decidim/home/application_mailer.rb
@@ -1,8 +1,0 @@
-module Decidim
-  module Home
-    class ApplicationMailer < ActionMailer::Base
-      default from: 'from@example.com'
-      layout 'mailer'
-    end
-  end
-end
