--- decidim-home/app/models/decidim/home/application_record.rb
@@ -1,7 +1,0 @@
-module Decidim
-  module Home
-    class ApplicationRecord < ActiveRecord::Base
-      self.abstract_class = true
-    end
-  end
-end
