+++ decidim-home/app/views/layouts/_footer.html.erb
@@ -1,19 +1,19 @@
 <div class="mini-footer">
-    <div class="row">
-        <div class="fwk-row">
-            <div class="fwk-col-md-2 fwk-col-sm-12">
-                <a rel="license" class="cc-badge part-logo-footer-gencat" href="http://www.gencat.cat" target="_blank">
-                    <%=image_tag("logo_generalitat_gris.png")%>
-                </a>
-            </div>
-            <div class="fwk-col-md-10 fwk-col-sm-12">
-                <div class="part-justify part-avis-legal">
-                    <a href="http://web.gencat.cat/<%= I18n.locale %>/menu-ajuda/ajuda/avis_legal/" target="_blank">
-                        <%= t("static.footer.legal_advice_title")%>:
-                    </a> 
-                    <%= t("static.footer.legal_advice_text")%>
-                </div>
-            </div>
-        </div>
-    </div>
+	<div class="row">
+		<div class="fwk-row">
+			<div class="fwk-col-md-2 fwk-col-sm-12">
+				<a rel="license" class="cc-badge part-logo-footer-gencat" href="http://www.gencat.cat" target="_blank">
+						<%=image_tag("logo_generalitat_gris.png")%>
+				</a>
+			</div>
+			<div class="fwk-col-md-10 fwk-col-sm-12">
+				<div class="part-justify part-avis-legal">
+					<a href="http://web.gencat.cat/<%= I18n.locale %>/menu-ajuda/ajuda/avis_legal/" target="_blank">
+							<%= t("static.footer.legal_advice_title")%>:
+					</a> 
+					<%= t("static.footer.legal_advice_text")%>
+				</div>
+			</div>
+		</div>
+	</div>
 </div>
