+++ decidim-home/app/views/layouts/decidim/_topNavbar.html.erb
@@ -1,22 +1,19 @@
 <div class="row small-up-3 part-navbar-background">
-    <div class="column">
-        <div class="part-topNavbar-content">
-            <svg aria-label="Navegació" role="img" class="icon--menu icon part-icon-bars" data-close>
-                <use xlink:href="/assets/decidim/icons-aa71fb7b9fb5d043b71aef057f4425f0f67c6e3642210bc7d2a80d6db01e4206.svg#icon-menu" class="part-icon-bars"></use>
-            </svg>
-        </div>
-    </div>
-    <div class="column">
-       
-         <div class="title-sm show-for-small">
-                            
-                                <%= link_to  "/", class: "part-a-white" do %>
-                                  Participa
-                                <% end %>
-                    </div>
-    </div>
+	<div class="column">
+		<div class="part-topNavbar-content">
+			<svg aria-label="Navegació" role="img" class="icon--menu icon part-icon-bars" data-close>
+				<use xlink:href="/assets/decidim/icons-aa71fb7b9fb5d043b71aef057f4425f0f67c6e3642210bc7d2a80d6db01e4206.svg#icon-menu" class="part-icon-bars"></use>
+			</svg>
+		</div>
+	</div>
+	<div class="column">
+		<div class="title-sm show-for-small">								
+			<%= link_to  "/", class: "part-a-white" do %>
+				Participa
+			<% end %>
+		</div>
+	</div>
 </div>
 <div class="row part-background-grey-navbar" title="Generalitat de Catalunya" onclick="window.open('http://web.gencat.cat/<%= I18n.locale %>/inici/','_blank');">
-    <div class="part-topNavbar-logo">
-    </div>
+	<div class="part-topNavbar-logo"></div>
 </div>
