--- decidim-home/app/views/pages/home.html.erb
@@ -1,40 +1,0 @@
-<% provide(:meta_title, current_organization.name) %>
-
-<%= render partial: 'pages/home/hero' %>
-
-<% if !translated_attribute(current_organization.description).blank? %>
-  <%= render partial: 'pages/home/sub_hero' %>
-<% end %>
-
-<% if false %>
-  <% found = false %>
-  <% searchProcesses.each do |process| %>
-    <% if found %>
-        <% break %>
-    <% end %>
-    <% searchFeatures(process.id).each do |feature| %>
-      <% if found %>
-        <% break %>
-        <% end %>
-      <% searchMeetings(feature.id).each_with_index do |meeting, i| %>
-        <% found = true %>
-        <% if found == true %>
-          <%= render partial: 'pages/home/next_meetings' %>
-        <% end %>
-        <% break %>
-      <% end %>
-    <% end %>
-  <% end %>
-<% end %>
-
-<%= render partial: 'pages/home/highlighted_processes' %>
-<%= render partial: 'pages/home/stable_areas' %>
-<%= render partial: 'pages/home/regulatory_inquiries' %>
-
-<%= render partial: 'pages/home/extended' %>
-
-<% if current_organization.show_statistics? %>
-  <%= render partial: 'pages/home/statistics' %>
-<% end %>
-
-<%= render partial: 'pages/home/footer_sub_hero' %>
