--- decidim-home/config/routes.rb
@@ -1,2 +1,0 @@
-Decidim::Home::Engine.routes.draw do
-end
