--- decidim-home/lib/tasks/decidim/home_tasks.rake
@@ -1,4 +1,0 @@
-# desc "Explaining what the task does"
-# task :decidim_home do
-#   # Task goes here
-# end
