+++ decidim-process-extended/app/queries/decidim/participatory_processes/organization_published_participatory_processes.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ParticipatoryProcesses
+    # This query class filters published processes given an organization.
+    class OrganizationPublishedParticipatoryProcesses < Rectify::Query
+      def initialize(organization)
+        @organization = organization
+      end
+
+      def query
+        Rectify::Query.merge(
+          OrganizationParticipatoryProcesses.new(@organization),
+          PublishedParticipatoryProcesses.new,
+          NotRegulation.new
+        ).query
+      end
+    end
+  end
+end
