+++ decidim-regulations/app/views/decidim/regulations/pages/home/_highlighted_regulations.html.erb
@@ -1,0 +1,35 @@
+<div class="row" id="highlighted-processes">
+  <h3 class="section-heading"><%= t(".active_processes") %></h3>
+  <div class="row collapse">
+    <div class="row small-up-1 smallmedium-up-2 mediumlarge-up-3
+      large-up-4 card-grid">
+      <% highlighted_regulations.each do |regulation| %>
+        <div class="column">
+          <article class="card card--process card--mini">
+            <%= link_to decidim_participatory_processes.participatory_process_path(regulation), class: "card__link" do %>
+              <div class="card__image-top"
+                style="background-image:url(<%= regulation.hero_image.url %>)"></div>
+            <% end %>
+            <div class="card__content">
+              <%= link_to decidim_participatory_processes.participatory_process_path(regulation), class: "card__link" do %>
+                <h4 class="card__title"><%= translated_attribute regulation.title %></h4>
+              <% end %>
+              <% if regulation.active_step %>
+                <span class="card--process__small">
+                  <%= t(".active_step") %>
+                  <strong><%= translated_attribute regulation.active_step.title %></strong>
+                </span>
+              <% end %>
+            </div>
+          </article>
+        </div>
+      <% end %>
+    </div>
+  </div>
+</div>
+<div class="row" id="see-all-processes">
+  <div class="columns small-centered small-12
+    smallmedium-8 medium-6 large-4">
+    <%= link_to t(".see_all_processes"), "regulations", class: "button expanded hollow button--sc home-section__cta" %>
+  </div>
+</div>
