+++ decidim-regulations/config/locales/ca.yml
@@ -3,6 +3,12 @@ ca:
     regulations_static: Participació en normativa
   decidim:
     regulations:
+      pages:
+        home:
+          highlighted_regulations:
+            active_processes: Participació en normativa
+            active_step: Pas actiu
+            see_all_processes: Veure totes les participacions
       regulation:
         show:
           email: Bústia de suport
